Change license date and name.
After a bit of thought, even if this isn't the correct license, let's at
least put the correct date and name.  The original of this file was
created automatically by GitHub, so had my name and the current year.

Yes, I know, Miffy Fluffy is just a Second Life account name, not a
legal person, but that's the name he put on it originally.